Core/Fix: Added Source TrinityCore Legion

+++ src/server/game/Grids/Grid.h
@@ -1,5 +1,6 @@
 /*
- * Copyright (C) 2016 DeathCore <http://www.noffearrdeathproject.org/>
+ * Copyright (C) 2008-2017 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -103,7 +104,7 @@ class Grid
         template<class T>
         uint32 GetWorldObjectCountInGrid() const
         {
-            return i_objects.template Count<T>();
+            return uint32(i_objects.template Count<T>());
         }
 
         /** Inserts a container type object into the grid.
@@ -140,4 +141,3 @@ class Grid
         //ActiveGridObjects m_activeGridObjects;
 };
 #endif
-
